AMP-13911
git-svn-id: file:///svn/branches/AMP_2_2_RELEASE@19091 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/content/util/DbUtil.java
@@ -155,8 +155,11 @@ public class DbUtil {
 					.iterator();
 			while (itr.hasNext()) {
                 AmpContentItemThumbnail delItem = itr.next();
-                Query q = sess.createQuery("delete from " + AmpContentItemThumbnail.class.getName() + " where ampContentItemThumbnailId = " + String.valueOf(delItem.getAmpContentItemThumbnailId().longValue()));
-                q.executeUpdate();
+                                
+                if (delItem.getAmpContentItemThumbnailId() != null){
+	                Query q = sess.createQuery("delete from " + AmpContentItemThumbnail.class.getName() + " where ampContentItemThumbnailId = " + String.valueOf(delItem.getAmpContentItemThumbnailId().longValue()));
+	                q.executeUpdate();
+                }
 				//sess.delete(delItem);
 			}
 			//tx.commit();
@@ -164,12 +167,12 @@ public class DbUtil {
 			if (e instanceof JDBCException)
 				throw (JDBCException) e;
 			logger.error("Exception " + e.toString());
-			try {
-				tx.rollback();
-			} catch (HibernateException ex) {
-				logger.error("rollback() failed");
-				logger.error(ex.toString());
-			}
+//			try {
+//				tx.rollback();
+//			} catch (HibernateException ex) {
+//				logger.error("rollback() failed");
+//				logger.error(ex.toString());
+//			}
 		} 
 
 	}
